refactor: 메인 캘린더 로직에서 캘린더 이벤트 로직 커스텀 훅으로 리팩토링
